Added WindowsAzure.Storage NuGet package,
because one of the first tasks of this microservice is going to be pulling
messages off e queue. Here at the beginning, I'm going to write this
microservice as a monolithic, untested microservice, because the initial
logic to get it up and running, is really going to be trivial.

We can always decouple it once we know a bit more about the desired axes
of variability.